phase2 ud slice->u8 in move table

+++ src/movetable.rs
@@ -599,7 +599,7 @@ impl MoveTable {
 
     pub fn get_next_phase2_ud_slice_coord(
         &self,
-        phase2_ud_slice_coord: u16,
+        phase2_ud_slice_coord: u8,
         move_action: u8,
     ) -> u8 {
         Self::get_next_coord(
